test: move the raycast tests and bench next to the transforms
